django_ml_regression_incomplete/requirements.txt
absl-py
astor
Django
django-crispy-forms
djangorestframework
gast==0.3.3
google-pasta
grpcio==1.32.0
gunicorn
h5py
joblib
Keras
Keras-Applications
Keras-Preprocessing
Markdown
numpy
pandas
protobuf
psycopg2-binary
python-dateutil
pytz
PyYAML
scikit-learn
scipy
six
sklearn
sqlparse
tensorboard
tensorflow
tensorflow-estimator
termcolor
Werkzeug
whitenoise
wrapt==1.12.1
xlrd
